+++ cupy/_math/arithmetic.py
