+++ cupy/_math/misc.py
