+++ cupy/_math/rational.py
