+++ cupy/_math/trigonometric.py
@@ -2,8 +2,8 @@ import numpy
 
 import cupy
 from cupy import core
-from cupy.math import sumprod
-from cupy.math import ufunc
+from cupy._math import sumprod
+from cupy._math import ufunc
 
 
 sin = ufunc.create_math_ufunc(
