+++ cupy/_math/ufunc.py
