+++ cupy/_math/window.py
