--- cupy/math/__init__.py
@@ -1,14 +1,0 @@
-# Functions from the following NumPy document
-# https://docs.scipy.org/doc/numpy/reference/routines.math.html
-
-# "NOQA" to suppress flake8 warning
-from cupy.math import arithmetic  # NOQA
-from cupy.math import explog  # NOQA
-from cupy.math import floating  # NOQA
-from cupy.math import hyperbolic  # NOQA
-from cupy.math import misc  # NOQA
-from cupy.math import rounding  # NOQA
-from cupy.math import special  # NOQA
-from cupy.math import sumprod  # NOQA
-from cupy.math import trigonometric  # NOQA
-from cupy.math import rational  # NOQA
